Major refactoring: - project recreation - fixed bug with glwindow destroying itself - major renaming in entity systems - major simplification of entity systems, including * introduction of unsafe type collections * removal of delayed messages etc - separated headers, luabind is no longer included everywhere - luabind::object changed to std::function and bind_callbacks introduced to set callbacks to these - added "glerr" after every gl call to check for errors - memory pool can dynamically destruct/move construct having just a type hash

+++ game_framework/systems/destroy_system.cpp
@@ -1,4 +1,3 @@
-#include "stdafx.h"
 #include <algorithm>
 #include <functional>
 
@@ -42,7 +41,7 @@ void destroy_system::consume_events(world& owner) {
 
 	owner.get_message_queue<messages::destroy_message>().clear();
 
-	for (int i = 0; i < to_destroy.size(); ++i) {
+	for (size_t i = 0; i < to_destroy.size(); ++i) {
 		owner.delete_entity(to_destroy[i].first);// , to_destroy[i].second);
 	}
 }
